Fixed report dialog resetting
